Section: ACP

+++ admin/template/searchuser.php
@@ -61,19 +61,20 @@
             <th><?php echo $tl["user"]["u2"]; ?></th>
             <th><?php echo $tl["menu"]["m9"]; ?></th>
             <th>
-              <button type="submit" name="lock" id="button_lock" class="btn btn-default btn-xs"><i
-                  class="fa fa-lock"></i></button>
+              <button type="submit" name="lock" id="button_lock" class="btn btn-default btn-xs">
+                <i class="fa fa-lock"></i>
+              </button>
             </th>
             <th>
-              <button type="submit" name="password" id="button_key" class="btn btn-default btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i
-                  class="fa fa-key"></i></button>
+              <button type="submit" name="password" id="button_key" class="btn btn-default btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;">
+                <i class="fa fa-key"></i>
+              </button>
             </th>
             <th></th>
             <th>
-              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                  class="fa fa-trash-o"></i></button>
+              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;" disabled="disabled">
+                <i class="fa fa-trash-o"></i>
+              </button>
             </th>
           </tr>
           </thead>
@@ -87,25 +88,32 @@
                 </td>
                 <td><?php echo $v["email"]; ?></td>
                 <td><?php echo $v["username"]; ?></td>
-                <td><?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $y) {
-                    if ($v["usergroupid"] == $y["id"]) { ?><a
-                      href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $y["id"]; ?>"><?php echo $y["name"]; ?></a><?php }
-                  } ?></td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>"><i
-                      class="fa fa-<?php if ($v["access"] == '1') { ?>check<?php } else { ?>lock<?php } ?>"></i></a>
+                <td>
+                  <?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $y) {
+                    if ($v["usergroupid"] == $y["id"]) { ?>
+                      <a href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $y["id"]; ?>"><?php echo $y["name"]; ?></a>
+                  <?php } } ?>
                 </td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>"
-                       onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i
-                      class="fa fa-key"></i></a></td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><i
-                      class="fa fa-edit"></i></a></td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>"
-                       onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                      class="fa fa-trash-o"></i></a></td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>">
+                    <i class="fa fa-<?php if ($v["access"] == '1') { ?>check<?php } else { ?>lock<?php } ?>"></i>
+                  </a>
+                </td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;">
+                    <i class="fa fa-key"></i>
+                  </a>
+                </td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>">
+                    <i class="fa fa-edit"></i>
+                  </a>
+                </td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;">
+                    <i class="fa fa-trash-o"></i>
+                  </a>
+                </td>
               </tr>
             <?php }
           } ?>
@@ -123,21 +131,26 @@
                     if ($v["usergroupid"] == $y["id"]) { ?><a
                       href="index.php?p=usergroup&amp;sp=user&amp;ssp=<?php echo $y["id"]; ?>"><?php echo $y["name"]; ?></a><?php }
                   } ?></td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>"><i
-                      class="fa fa-<?php if ($v["access"] == '1') { ?>check<?php } else { ?>lock<?php } ?>"></i></a>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=lock&amp;ssp=<?php echo $v["id"]; ?>">
+                    <i class="fa fa-<?php if ($v["access"] == '1') { ?>check<?php } else { ?>lock<?php } ?>"></i>
+                  </a>
                 </td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>"
-                       onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;"><i
-                      class="fa fa-key"></i></a></td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>"><i
-                      class="fa fa-edit"></i></a></td>
-                <td><a class="btn btn-default btn-xs"
-                       href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>"
-                       onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                      class="fa fa-trash-o"></i></a></td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=password&amp;ssp=<?php echo $v["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["pass"]; ?>'))return false;">
+                    <i class="fa fa-key"></i>
+                  </a>
+                </td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=edit&amp;ssp=<?php echo $v["id"]; ?>">
+                    <i class="fa fa-edit"></i>
+                  </a>
+                </td>
+                <td>
+                  <a class="btn btn-default btn-xs" href="index.php?p=user&amp;sp=delete&amp;ssp=<?php echo $v["id"]; ?>" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;">
+                    <i class="fa fa-trash-o"></i>
+                  </a>
+                </td>
               </tr>
             <?php }
           } ?>
@@ -165,16 +178,27 @@
     <i title="<?php echo $tl["icons"]["i1"]; ?>" class="fa fa-trash-o"></i>
   </div>
 
-  <!-- JavaScript for select all -->
-
   <script type="text/javascript">
     $(document).ready(function () {
+
+      /* Check all checkbox */
       $("#jak_delete_all").click(function () {
         var checked_status = this.checked;
         $(".highlight").each(function () {
           this.checked = checked_status;
         });
       });
+
+      /* Disable submit button if checkbox is not checked */
+      var the_terms = $('.highlight');
+      the_terms.click(function() {
+        if ($(this).is(":checked")) {
+          $("#button_delete").removeAttr("disabled");
+        } else {
+          $("#button_delete").attr("disabled", "disabled");
+        }
+      });
+
     });
   </script>
